+++ src/instanceConfigFields.ts
@@ -1,13 +1,16 @@
 import { IntegrationInstanceConfigFieldMap } from '@jupiterone/integration-sdk-core';
+import { GitlabIntegrationConfig } from './types';
 
-const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
+const instanceConfigFields: IntegrationInstanceConfigFieldMap<GitlabIntegrationConfig> = {
   personalToken: {
     type: 'string',
     mask: true,
   },
   baseUrl: {
     type: 'string',
-    mask: true,
+  },
+  mergeRequestsUpdatedAfter: {
+    type: 'string',
   },
 };