+++ test/schemas.ts
@@ -93,5 +93,10 @@ export const prSchema: GraphObjectSchema = {
 
     createdOn: { type: 'number' },
     webLink: { type: 'string' },
+    mergedAt: { type: 'number' },
+    closedAt: { type: 'number' },
+    sha: { type: 'string' },
+    mergeCommitSha: { type: 'string' },
+    squashCommitSha: { type: 'string' },
   },
 };